Update and rename DBContext.cs to TOTDBContext.cs
